Add time view with 12min slices
